backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
cx-Oracle==8.3.0
pandas==2.2.0
numpy==1.26.3
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
python-dotenv==1.0.0
alembic==1.13.1
